LIBCIR-295. Implement Random DOI minting
Restored the following UMD-customized files, updating them to match
their DSpace 7.5 counterparts, while preserving and marking the UMD
customizations:

* dspace/modules/additions/src/main/java/org/dspace/identifier/DOIIdentifierProvider.java
* dspace/modules/additions/src/main/java/org/dspace/identifier/doi/DataCiteConnector.java
* dspace/modules/additions/src/main/java/org/dspace/identifier/doi/DOIOrganiser.java

Wrote up information about DOI customizations in
“dspace/docs/MdsoarDOI.md”.

One issue was that DSpace 7, unlike DSpace 6, does not ensure that the
Java jar files are loaded in a specific order when running the “dspace”
script. This was causing issues when running the
“/dspace/bin/dspace doi-organiser” command, because the
“dspace-api.jar” file would be in the classpath before the
“additions.jar” file, resulting in the stock DSpace classes being used
instead of the UMD-customized versions. This is a known issue, as
evidenced by DSpace issues 3248 and 8506 (and so may be fixed in a
future release). Until then, modified the “dspace/bin/dspace” script to
place the “additions.jar” file before any of the standard DSpace jar
files in the classpath.

Added "DOI Configuration" section to the "local.cfg.EXAMPLE" file.

Modified "identifier-service.xml" to support parameterizing the
DataCite host via a the "identifier.doi.datacite.host" property file.

https://umd-dit.atlassian.net/browse/LIBCIR-295