LIBCIR-326. webui.browse.link CrossLinks: Fix for multiple exact matches
Fixes an issue in the "CrossLinks" class to enable multiple
"webui.browse.link" configuration entries with exact matches to
different indexes to be processed properly.

This fixes a stock DSpace bug. The associated DSpace issue is

https://github.com/DSpace/DSpace/issues/8935

and a change similar to this one (except in the "dspace-api" module)
was submitted in a pull request:

https://github.com/DSpace/DSpace/pull/8936

https://umd-dit.atlassian.net/browse/LIBCIR-326