LIBCIR-322. Added CrossLinks configuration for MD-SOAR
Added configuration to "local.cfg.EXAMPLE" to configure "crosslinks"
for the simple item page, so that specific fields will be hyperlinked
to their appropriate "browse by" search page.

Added "SimpleItemPage.md" to record MD-SOAR customizations to the
"simple item" page.

Updated "MdsoarCustomizations.md" to link to the "SimpleItemPage.md" and
rearranged the links to be in alphabetical order.

https://umd-dit.atlassian.net/browse/LIBCIR-322