LIBCIR-338. Disable "buildOnCommit" for Solr suggesters
Having Solr "buildOnCommit" enabled, so that the suggester indexes are
rebuilt whenever a Solr item is committed, severely impacts the DSpace
submission form, as the form appears to trigger a Solr commit when the
form is autosaved, saved by the user, or deposited. It also makes file
uploads much slower.

The downside of disable "buildOnCommit" is that new entries made by
users to the
"Format" and "Subject Keywords" fields will not be added to suggester
indexes until either Solr restarts, or a cron job is run to trigger
a rebuild.

Leaving "buildOnStatup" enabled because it does not appear to greatly
increase the startup time of Solr. If this becomes an issue, it could
possibly be disabled as well.

https://umd-dit.atlassian.net/browse/LIBCIR-338