LIBCIR-338. Added "umd-solr-rebuild-suggesters" DSpace task
Added a "umd-solr-rebuild-suggesters" DSpace script to trigger a
rebuild of all the Solr suggester indexes.

This can be triggered manually by running

```bash
$ /dspace/bin/dspace umd-solr-rebuild-suggesters
```

but more typically it will be run by cron.

Pulled the "/suggest" request handler into a public constant in
"SolrSuggestService", so it would be available to the
"UmdSolrRebuildSuggestersClient" class.

Updated the "dspace/docs/SolrAutosuggest.md" documentation with
information about the "umd-solr-rebuild-suggesters" DSpace script.

https://umd-dit.atlassian.net/browse/LIBCIR-338