LIBCIR-340. Enable optimization for "select-collection-step" query
Enabling the "org.dspace.content.Collection.findAuthorizedPerformanceOptimize"
property significantly speeds up the display of the submission form
for non-admins when selecting the “Submit item to MD-SOAR” link on
the home page.

This property is not enabled by default (as outlined in
https://github.com/DSpace/DSpace/issues/6033) because it is not
compatible with Shibboleth or LDAP authentication and special groups.
(see also the note for this property in "dspace/config/dspace.config".

Since MD-SOAR does not use Shibboeth or LDAP authentication, and
has a large number of collection, it seems to meet the criteria for
enabling this property.

https://umd-dit.atlassian.net/browse/LIBCIR-340