LIBCIR-352. Enable all users to view usage statistics
Restores the stock DSpace behavior of allowing all users (including
anonymous users) to view the usage statistics.

This reverses an incorrect configuration change made to MD-SOAR in
LIBCIR-299, which restricted the usage statistics to admins.

Also removed the property from the list of customized properties in
"dspace/docs/MdsoarCustomizations.md"

https://umd-dit.atlassian.net/browse/LIBCIR-352