LIBCIR-368. Removed customized CrossLinks functionality
Removed "CrossLinks.java" and "CrossLinksTest.java" from the
"dspace/additions" module because these files were added to fix
a bug in DSpace 7.5.

The change was submitted to the upstream DSpace, and accepted into the
stock DSpace files, so these replacements for the stock files are no
longer needed.

See https://github.com/DSpace/DSpace/issues/8935 and
https://github.com/DSpace/DSpace/pull/8936

https://umd-dit.atlassian.net/browse/LIBCIR-368