LIBCIR-386. Fix OpenSearch NullPointerException for unknown valid UUIDs
Fixes a NullPointerException when the "scope" parameter provided to
the OpenSearch endpoint is a valid UUID, but is not a UUID associated
with a Community or Collection.

Instead of throwing a NullPointerException, this change modifies the
code to return a null scope (resulting in an "unscoped" OpenSearch
request), as if the UUID were invalid, or otherwise not usable.

This fix was provided to DSpace as Pull Request #9482 -
<https://github.com/DSpace/DSpace/pull/9482>. The customization markers
added in this issue can be removed when MD-SOAR is updated to a DSpace
version containing this change.

https://umd-dit.atlassian.net/browse/LIBCIR-386