LIBCIR-430. Fix checkpointing for "dspace checker" command
Fixes the incremental storage of checksums for the "dspace checker"
command by actually committing the changes to the database.

This change was provided to DSpace as DSpace Pull Request 10508.

Replacing "uncacheEntity" with "commit" as testing has shown that this
is necessary to ensure that the checksums are properly updated when the
checker is interrupted.

The provided integration test fails if checksums checked before the
interruption are not properly recorded in the database. This class
is slightly different from the version provided to DSpace for DSpace 9,
due to changes in the use of the "java.util.Date" and
"java.time.Instant" classes.

https://umd-dit.atlassian.net/browse/LIBCIR-430