LIBCIR-438. Comment out "codecov" GitHub Action
Commented out the "codecov" GitHub action that is part of the build
process because UMD does not have an appropriate key for uploading
the results to codecov.io.

Commenting out the action enables the builds to be marked as
successful (as long as they pass the unit and integration tests).

https://umd-dit.atlassian.net/browse/LIBCIR-438