LIBCIR-451. Replace Author/Contributor/Advisor hint text with I18n keys
The MD-SOAR Community Admins have requested changes to the "hints"
for the "Author(s)", "Contributor(s)", and "Advisor(s)" fields on the
submission form to make more prominent that the names should be in
"Last Name, First Name" format. The requested changes include using
**bold** text in the hint.

Any HTML tags added directly to the hint text are displayed AS-IS in
GUI, instead of being interpreted by the browser. To enable HTML tags to
be included in the hints, replaced the hard-coded text with I18n keys,
that can then be used with Angular's "translate" module to retrieve the
text from an Angular front-end localization file (such as
“src/assets/i18n/en.json5”). Text retrieved from a localization file
can use a limited subset of HTML tags (include "<b>" and "<strong>").

https://umd-dit.atlassian.net/browse/LIBCIR-451