LIBCIR-458. Fix bitstream path/assetstore path comparison
Modified "DSBitstoreService" to use "getCanonicalPath" to compare the
bitstream path to the assetstore path.

This is necessary for the
"org.dspace.storage.bitstore.BitstreamStorageServiceImplIT" tests to
pass on macOS, because the temporary directory used by those tests
contains a symbolic link.

This change has been provided to DSpace in pull request 11347. See
that pull request and the associate issue (11346) for additional
information.

https://umd-dit.atlassian.net/browse/LIBCIR-458