LIBCIR-458. Removed BitstreamStorageServiceImpl change for AWS migration
Removed a UMD customization that was made to support solely to the AWS
migration.

When performing the migration, the DSpace code would fail when run in
Kubernetes sandbox/test/qa because the assetstores for those
environments only had a subset of the production data. The DSpace
code assumes that no assets will be missing in the migration to S3, and
the UMD customization allowed missing files to be skipped.

Now that the AWS migration is complete, this customization is no longer
necessary.

https://umd-dit.atlassian.net/browse/LIBCIR-458