LIBCIR-457. Add script to prepare and run postfix.
Port of DRUM changes from LIBDEVOPS-83 into MD-SOAR to support sending
emails from the "cron" container that do not go through the DSpace
email handler.

This commit properly configures the Postfix mail handler with a
script to:

* Run Postfix to run in a chroot environment.
* Fix directory ownership and permissions
* Start the Postfix service

Further changes in the k8s-mdsoar repository ensure that Postfix is
started as part of the "cron" container.

https://umd-dit.atlassian.net/browse/LIBCIR-457